Crate rustc_hex

Hex binary-to-text encoding

Re-exports

pub use self::FromHexError::*;

Structs

FromHexIter

An iterator decoding hex-encoded characters into bytes.

ToHexIter

An iterator converting byte slice to a set of hex characters.

Enums

FromHexError

Errors that can occur when decoding a hex encoded string

Traits

FromHex

A from-hex conversion trait.

ToHex

A trait for converting a value to hexadecimal encoding