Trait rustc_hex::ToHex

pub trait ToHex {
    fn to_hex<T: FromIterator<char>>(&self) -> T;
}

A trait for converting a value to hexadecimal encoding

Required Methods

fn to_hex<T: FromIterator<char>>(&self) -> T

Converts the value of self to a hex value, constructed from an iterator of characters.